Rename document_inspect -> doc_inspect
